Fixed check errors

+++ golang/main.go
@@ -22,13 +22,13 @@ func func_resolution(x float64) float64 {
 	square := square(x)
 	diff := square - 2.5
 
-	mod := math.Abs(float64(diff))
+	mod := math.Abs(diff)
 
-	root4 := math.Pow(float64(mod), root_nth_degree(4))
+	root4 := math.Pow(mod, root_nth_degree(4))
 
 	nat_log := math.Log(square)
 
-	root3 := math.Pow(float64(nat_log), root_nth_degree(3))
+	root3 := math.Pow(nat_log, root_nth_degree(3))
 
 	y := root4 + root3
